backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

sqlmodel==0.0.22
psycopg2-binary==2.9.10
alembic==1.13.1

pydantic==2.10.3
pydantic-settings==2.7.1

openai>=1.57.4
httpx==0.27.0

python-multipart==0.0.20
python-dotenv==1.0.1

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

slowapi>=0.1.9
python-json-logger>=2.0.7
sse-starlette==1.6.5





